pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gauntlet-eval"
version = "0.1.0"
description = "Adversarial multi-agent eval harness for Claude API pipelines"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["ai", "agents", "evaluation", "llm", "claude", "adversarial"]
dependencies = [
    "anthropic>=0.25.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "httpx>=0.27", "ruff>=0.4"]

[project.scripts]
gauntlet = "gauntlet.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["gauntlet*"]
